AI-features/requirements.txt
google-generativeai
gradio
langchain-text-splitters
pymupdf
pdf2image
google-generativeai
poppler-utils
pypdf
pymupdf
IPython
langchain_community
langchain-text-splitters
chromadb==0.5.0